refactroing search
